If StringLiteral meta-data already includes a raw form, use that.
This is only valid if the `node.value` has not changed from the original value.
This also addresses a strange condition in babylon where the following code:

  "\u2028";

Will be parsed like so:

  {
    type: "StringLiteral",
    value: "",
    extra: {
      rawValue: "",
      raw: "\"\\u2028\""
    }
  }

Note that the node's value is an empty string, so that node can't be correctly
rendered without the original `extra.raw` value.